- fix recursive import by moving a funtion out of util - unit tests
